Changed paths to always be relative to the cwd.
The only reason where this fails is, if the cwd is not inside the project
directory AND the vcs does not support absolute file names. I am not sure if
people do this in reality, but if it actually is a problem it should be solved
inside of repo_detect and only for absolute paths.